requirements.txt
python-telegram-bot>=20.7
SpeechRecognition>=3.10.0
openai>=1.3.0
python-dotenv>=1.0.0
pydub>=0.25.1
Levenshtein>=0.21.1

